% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{model_run}
\alias{model_run}
\title{Run the QRISK3 model}
\usage{
model_run(model_input = NULL, hist = FALSE)
}
\arguments{
\item{model_input}{Named list or data frame of patient inputs.}
}
\value{
Data frame with QRISK3 risk score columns.
}
\description{
Run the QRISK3 model
}
